Scraper/VPR_Builder/requirements.txt
numpy>=1.24.0
faiss-cpu>=1.7.4
boto3>=1.34.0
requests>=2.31.0
